+++ scripts/src/blocks/OutputBlock.ts
@@ -19,14 +19,16 @@ class OutputBlock extends Block {
 
             if(this.isVariable) {
                 console.log(message + globalVariables.get(this.variable));
-                const msg = new messageBox(message + globalVariables.get(this.variable));
+                const msg = new MessageBox(message + globalVariables.get(this.variable));
                 msg.okBtn.addEventListener("click", () => this.executeNextBlock());
             }
             else {
                 console.log(message);
-                const msg = new messageBox(message);
+                const msg = new MessageBox(message);
                 msg.okBtn.addEventListener("click", () => this.executeNextBlock());
             }
+
+            
         }, runSpeed/5);
     }